armira-master/requirements.txt
Flask
numpy
opencv-python
Pillow
openai
python-dotenv
requests
